Factory Satellite fixtures to ignore in upgrade tests
