+++ conftest.py
@@ -17,6 +17,7 @@ pytest_plugins = [
     'pytest_plugins.rerun_rp.rerun_rp',
     'pytest_plugins.fspath_plugins',
     'pytest_plugins.fixture_collection',
+    'pytest_plugins.factory_collection',
     # Fixtures
     'pytest_fixtures.core.broker',
     'pytest_fixtures.core.contenthosts',
