+++ pytest_fixtures/core/broker.py
@@ -117,126 +117,6 @@ def capsule_factory():
     return factory
 
 
-@pytest.fixture
-def satellite_host(satellite_factory):
-    """A fixture that provides a Satellite based on config settings"""
-    new_sat = satellite_factory()
-    yield new_sat
-    new_sat.teardown()
-    Broker(hosts=[new_sat]).checkin()
-
-
-@pytest.fixture(scope='module')
-def module_satellite_host(satellite_factory):
-    """A fixture that provides a Satellite based on config settings"""
-    new_sat = satellite_factory()
-    yield new_sat
-    new_sat.teardown()
-    Broker(hosts=[new_sat]).checkin()
-
-
-@pytest.fixture(scope='session')
-def session_satellite_host(satellite_factory):
-    """A fixture that provides a Satellite based on config settings"""
-    new_sat = satellite_factory()
-    yield new_sat
-    new_sat.teardown()
-    Broker(hosts=[new_sat]).checkin()
-
-
-@pytest.fixture
-def capsule_host(capsule_factory):
-    """A fixture that provides a Capsule based on config settings"""
-    new_cap = capsule_factory()
-    yield new_cap
-    new_cap.teardown()
-    Broker(hosts=[new_cap]).checkin()
-
-
-@pytest.fixture(scope='module')
-def module_capsule_host(capsule_factory):
-    """A fixture that provides a Capsule based on config settings"""
-    new_cap = capsule_factory()
-    yield new_cap
-    new_cap.teardown()
-    Broker(hosts=[new_cap]).checkin()
-
-
-@pytest.fixture(scope='session')
-def session_capsule_host(capsule_factory):
-    """A fixture that provides a Capsule based on config settings"""
-    new_cap = capsule_factory()
-    yield new_cap
-    new_cap.teardown()
-    Broker(hosts=[new_cap]).checkin()
-
-
-@pytest.fixture
-def capsule_configured(capsule_host, target_sat):
-    """Configure the capsule instance with the satellite from settings.server.hostname"""
-    capsule_host.capsule_setup(sat_host=target_sat)
-    yield capsule_host
-
-
-@pytest.fixture(scope='module')
-def module_capsule_configured(module_capsule_host, module_target_sat):
-    """Configure the capsule instance with the satellite from settings.server.hostname"""
-    module_capsule_host.capsule_setup(sat_host=module_target_sat)
-    yield module_capsule_host
-
-
-@pytest.fixture(scope='session')
-def session_capsule_configured(session_capsule_host, session_target_sat):
-    """Configure the capsule instance with the satellite from settings.server.hostname"""
-    session_capsule_host.capsule_setup(sat_host=session_target_sat)
-    yield session_capsule_host
-
-
-@pytest.fixture(scope='module')
-def module_capsule_configured_mqtt(module_capsule_configured):
-    """Configure the capsule instance with the satellite from settings.server.hostname,
-    enable MQTT broker"""
-    module_capsule_configured.set_rex_script_mode_provider('pull-mqtt')
-    result = module_capsule_configured.execute('systemctl status mosquitto')
-    assert result.status == 0, 'MQTT broker is not running'
-    result = module_capsule_configured.execute('firewall-cmd --permanent --add-port="1883/tcp"')
-    assert result.status == 0, 'Failed to open mqtt port on capsule'
-    module_capsule_configured.execute('firewall-cmd --reload')
-    yield module_capsule_configured
-
-
-@pytest.fixture(scope='module')
-def module_lb_capsule(retry_limit=3, delay=300, **broker_args):
-    """A fixture that spins 2 capsule for loadbalancer
-    :return: List of capsules
-    """
-    if settings.capsule.get('deploy_arguments'):
-        resolved = _resolve_deploy_args(settings.capsule.deploy_arguments)
-        settings.set('capsule.deploy_arguments', resolved)
-        broker_args.update(settings.capsule.deploy_arguments)
-        timeout = (1200 + delay) * retry_limit
-        hosts = Broker(
-            host_class=Capsule,
-            workflow=settings.capsule.deploy_workflow,
-            _count=2,
-            **broker_args,
-        )
-        cap_hosts = wait_for(hosts.checkout, timeout=timeout, delay=delay)
-
-    yield cap_hosts.out
-
-    _ = [cap.teardown() for cap in cap_hosts.out]
-    Broker(hosts=cap_hosts.out).checkin()
-
-
-@pytest.fixture(scope='module')
-def module_capsule_configured_async_ssh(module_capsule_configured):
-    """Configure the capsule instance with the satellite from settings.server.hostname,
-    enable MQTT broker"""
-    module_capsule_configured.set_rex_script_mode_provider('ssh-async')
-    yield module_capsule_configured
-
-
 @pytest.fixture(scope='module')
 def module_discovery_sat(
     module_provisioning_sat,
